+++ src/main/kotlin/me/aberrantfox/warmbot/commands/ConfigurationCommands.kt
@@ -50,4 +50,17 @@ fun configurationCommands(configuration: Configuration, persistenceService: Pers
             it.respond(Locale.inject({ SET_STAFF_ROLE_SUCCESSFUL }, "staffRoleName" to staffRole.name))
         }
     }
+
+    command("SetLoggingChannel") {
+        requiresGuild = true
+        description = "Set the channel where events will be logged."
+        expect(TextChannelArg)
+        execute {
+            val loggingChannel = it.args.component1() as TextChannel
+
+            configuration.getGuildConfig(it.message.guild.id)!!.loggingConfiguration.loggingChannel = loggingChannel.id
+            persistenceService.save(configuration)
+            it.respond("Successfully set the logging channel to ${loggingChannel.name}")
+        }
+    }
 }